Split up map component: one static map and one editable map
