+++ ui/src/design-system/map/types.ts
@@ -1,0 +1,1 @@
+export { LatLng as MarkerPosition, Map } from 'leaflet'
